DATA ANALYSIS USING PANDAS
The homes.csv dataset (available on the classnotes) shows attributes from a sample of residential
houses. Each row includes the attributes of a house (described on page 2). Use python library Pandas
to answer the following questions. For each one show the python command as well as the resulting
output.
1. Find the number of houses from each style.
2. Find the smallest, median, and largest value for each numerical column in the dataset.
3. Find the most expensive house with at least three bedrooms.
4. List all houses having two to four bedrooms with area exceeding 4000 square feet and price not
less than 350,000 dollars.
5. What is the average lot size of houses built after 1970?
6. On average how much more expensive are houses with a pool?
7. Use groupby() to find the lowest and largest prices of houses for each quality category.
8. Construct a cross tabulation table showing the number of houses classified by style (rows) and by
number of beds (columns).
9. Create a table showing all attributes of the least expensive style 1 house.
10. Create a DataFrame having all the categorical columns only. Then for each column find the
number of houses in each category.